# packages/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Project Booth API
  summary: Platform-neutral API contract for Project Booth clients.
  description: |
    Product endpoints use path versioning beginning with `/v1`. Operational
    liveness remains at the unversioned `/health/live` path so container
    orchestration does not depend on a product API version.

    Identity endpoints exchange provider credentials for Project Booth
    sessions. Game endpoints trust only the resulting internal account and
    never a platform-provider identifier.
  version: 0.3.0
...
servers:
  - url: "{scheme}://{authority}"
    description: Environment-selected API origin.
    variables:
      scheme:
        default: https
        enum:
          - https
          - http
        description: HTTPS in deployed environments; HTTP is local-only.
      authority:
        default: api.invalid
        description: Host and optional port supplied by the client environment.
...
security:
  - bearerAuth: []
paths:
...
  /v1/matches/{matchId}/snapshot:
    parameters:
      - $ref: "#/components/parameters/MatchId"
    get:
      tags:
        - Matches
      summary: Get an account-safe authoritative match snapshot
      operationId: getMatchSnapshot
      responses:
        "200":
          description: Snapshot and the recipient cursor captured with it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchSnapshot"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/realtime/events:
    get:
      tags:
        - Realtime
      summary: Resume the authenticated recipient event stream
      operationId: resumeRealtimeEvents
      parameters:
        - name: afterCursor
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: A strictly ordered page of recipient-safe events.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealtimeEventPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A short-lived Project Booth access token whose encoding is opaque to
        clients. Authentication behavior is implemented in a later milestone.
  parameters:
    MatchId:
      name: matchId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestId:
      description: Request trace identifier for support and diagnostics.
      schema:
        type: string
        minLength: 1
        maxLength: 128
...
    WWWAuthenticate:
      description: Bearer authentication challenge.
      schema:
        type: string
  schemas:
...
    MatchSnapshotRosterEntry:
      type: object
      additionalProperties: false
      required:
        - avatarKey
        - displayName
        - ready
        - status
        - userId
      properties:
        avatarKey:
          type: string
        displayName:
          type: string
        ready:
          type: boolean
        status:
          type: string
          enum:
            - active
            - eliminated
        userId:
          type: string
          format: uuid
    MatchSnapshot:
      type: object
      additionalProperties: false
      required:
        - lastRecipientCursor
        - matchId
        - matchVersion
        - phase
        - phaseDeadline
        - roster
        - ruleset
        - self
      properties:
        lastRecipientCursor:
          type: integer
          minimum: 0
        matchId:
          type: string
          format: uuid
        matchVersion:
          type: integer
          minimum: 1
        phase:
          type: string
        phaseDeadline:
          type:
            - string
            - "null"
          format: date-time
        roster:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/MatchSnapshotRosterEntry"
        ruleset:
          type: object
          additionalProperties: false
          required:
            - rulesetId
            - rulesetVersion
          properties:
            rulesetId:
              type: string
              format: uuid
            rulesetVersion:
              type: integer
              minimum: 1
        self:
          type: object
          additionalProperties: false
          required:
            - ready
            - userId
          properties:
            ready:
              type: boolean
            userId:
              type: string
              format: uuid
    RealtimeEvent:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - eventId
        - type
        - occurredAt
        - matchId
        - matchVersion
        - recipientCursor
        - audience
        - payload
      properties:
        schemaVersion:
          type: integer
          const: 1
        eventId:
          type: string
          format: uuid
        type:
          type: string
        occurredAt:
          type: string
          format: date-time
        matchId:
          type: string
          format: uuid
        matchVersion:
          type: integer
          minimum: 1
        recipientCursor:
          type: integer
          minimum: 1
        audience:
          type: string
          enum:
            - player
            - participants
            - active
        recipientUserId:
          type: string
          format: uuid
        payload:
          type: object
          additionalProperties: true
    RealtimeEventPage:
      type: object
      additionalProperties: false
      required:
        - events
        - hasMore
        - nextCursor
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/RealtimeEvent"
        hasMore:
          type: boolean
        nextCursor:
          type: integer
          minimum: 0
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
    ApiError:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - traceId
      properties:
        code:
          type: string
          pattern: ^[a-z][a-z0-9_]*$
          maxLength: 64
          description: Stable machine-readable error code.
        message:
          type: string
          minLength: 1
          maxLength: 500
          description: Safe human-readable explanation.
        traceId:
          type: string
          minLength: 1
          maxLength: 128
          description: Identifier supplied when contacting support.
        details:
          type: object
          additionalProperties: true
          description: Safe structured context specific to the error code.
...
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication is missing, expired, or invalid.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
        WWW-Authenticate:
          $ref: "#/components/headers/WWWAuthenticate"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: The authenticated account cannot perform this operation.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested account or profile does not exist.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Identity services are unavailable or not configured.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"